requirements.txt
streamlit 
langchain 
langchain-google-genai 
faiss-cpu 
PyPDF2
langchain-community
python-dotenv